+++ packages/use-dataloader/src/index.ts
@@ -4,3 +4,4 @@ export {
 } from './DataLoaderProvider'
 export { default as useDataLoader } from './useDataLoader'
 export { default as usePaginatedDataLoader } from './usePaginatedDataLoader'
+export type { UseDataLoaderConfig } from './types'