+++ packages/jest-helpers/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.18 (2022-09-20)
+
+**Note:** Version bump only for package @scaleway/jest-helpers
+
+
+
+
+
+## 2.0.17 (2022-09-12)
+
+**Note:** Version bump only for package @scaleway/jest-helpers
+
+
+
+
+
 ## 2.0.16 (2022-09-06)