+++ packages/use-dataloader/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.16.1 (2022-09-26)
+
+**Note:** Version bump only for package @scaleway/use-dataloader
+
+
+
+
+
 ## [2.16.0](https://github.com/scaleway/scaleway-lib/compare/@scaleway/use-dataloader@2.15.0...@scaleway/use-dataloader@2.16.0) (2022-08-19)